screen position to world position, position 값 float로
screen position to world position
- 터치 포지션 수정
 position 값 float
- 자연스럽고 정교한 움직임을 위해 int 형에서 float로 수정

+++ AirCom2us/Assets/NetworkUtils.cs
@@ -22,7 +22,7 @@ public static class NetworkUtils
     {
         Debug.Log("SendMovePacket - " + touchPos.x + ", " + touchPos.y);
 
-        cs_packet_move movePacket = new cs_packet_move((byte)Marshal.SizeOf(typeof(cs_packet_move)), Convert.ToByte(CS.MOVE), (int)touchPos.x, (int)touchPos.y, 0);
+        cs_packet_move movePacket = new cs_packet_move((byte)Marshal.SizeOf(typeof(cs_packet_move)), Convert.ToByte(CS.MOVE),touchPos.x,touchPos.y, 0);
 
         byte[] packet = new byte[1];
         StructToBytes(movePacket, ref packet);
